+++ CHANGES.md
@@ -1,8 +1,11 @@
 ## Changelog ##
+- [0.0.6]:
+    - Added first part of the question from catalog functionality.
+
 - [0.0.5]:
     - Added first part of quickstart functionality.
     - Added tabarea component for layout.
-    - Moved the fetching of PINGO data into seperat methods in an own class.
+    - Moved the fetching of PINGO data into separat methods in an own class.
 
 - [0.0.4]:
     - Added more functionality.